Arrumadas as dependências (main não depende de rio, grade sim)

+++ rio.h
@@ -1,5 +1,5 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void primeiraLinha(int *linha, int largura, float limiteDasMargens, int fluxoDesejado);
+void primeiraLinha(float *linha, int largura, float limiteDasMargens, int fluxoDesejado);
